Split Widgets.cs into one-class-per-file.
