+++ src/kOS/kOS.csproj
@@ -167,7 +167,16 @@
     <Compile Include="Suffixed\VesselEta.cs" />
     <Compile Include="Suffixed\WaypointValue.cs" />
     <Compile Include="Execution\TransferManager.cs" />
-    <Compile Include="Suffixed\Widgets.cs" />
+    <Compile Include="Suffixed\Widget\Slider.cs" />
+    <Compile Include="Suffixed\Widget\Spacing.cs" />
+    <Compile Include="Suffixed\Widget\Label.cs" />
+    <Compile Include="Suffixed\Widget\TextField.cs" />
+    <Compile Include="Suffixed\Widget\Button.cs" />
+    <Compile Include="Suffixed\Widget\PopupMenu.cs" />
+    <Compile Include="Suffixed\Widget\Box.cs" />
+    <Compile Include="Suffixed\Widget\ScrollBox.cs" />
+    <Compile Include="Suffixed\Widget\GUIWidgets.cs" />
+    <Compile Include="Suffixed\Widget\Widget.cs" />
     <Compile Include="UserIO\TelnetMainServer.cs" />
     <Compile Include="UserIO\TelnetSingletonServer.cs" />
     <Compile Include="UserIO\TelnetWelcomeMenu.cs" />
